requirements.txt
streamlit==1.39.0
pandas==2.2.2
snowflake-connector-python==3.10.1
python-dateutil
